Updated trade module with refactoring from core module
