+++ trade/src/main/java/co/bitsquared/btceparser/trade/data/ActiveOrder.java
@@ -1,7 +1,7 @@
 package co.bitsquared.btceparser.trade.data;
 
-import co.bitsquared.btceparser.core.DepthType;
-import co.bitsquared.btceparser.core.TradingPair;
+import co.bitsquared.btceparser.core.data.DepthType;
+import co.bitsquared.btceparser.core.data.TradingPair;
 import co.bitsquared.btceparser.core.currency.BaseCurrency;
 import org.json.JSONObject;
