+++ trade/src/main/java/co/bitsquared/btceparser/trade/requests/OrderInfoRequest.java
@@ -1,7 +1,7 @@
 package co.bitsquared.btceparser.trade.requests;
 
-import co.bitsquared.btceparser.core.TradingPair;
-import co.bitsquared.btceparser.core.DepthType;
+import co.bitsquared.btceparser.core.data.DepthType;
+import co.bitsquared.btceparser.core.data.TradingPair;
 import co.bitsquared.btceparser.trade.TAPI;
 import co.bitsquared.btceparser.trade.authentication.Authenticator;
 import co.bitsquared.btceparser.trade.callbacks.OrderInfoCallback;
